requirements.txt
discord.py
requests
jinja2
ruamel.yaml